+++ requirements.txt
@@ -12,8 +12,6 @@ chardet==3.0.4
 colorama==0.3.9
 decorator==4.3.0
 Django==2.1.2
-django-cors-headers==2.4.0
-django-debug-toolbar==1.10.1
 django-extensions==2.1.3
 django-filter==2.0.0
 django-nested-admin==3.1.2
@@ -55,7 +53,6 @@ sentry-sdk==0.5.5
 Serializer==0.2.1
 simplejson==3.16.0
 six==1.11.0
-sqlparse==0.2.4
 termcolor==1.1.0
 texttable==0.9.1
 traitlets==4.3.2
